feat: add full screen button
