+++ covid-dashboard/src/style.css
@@ -152,8 +152,25 @@ h1 {
   color: white
 }
 
+.button__expand {
+  outline: none;
+  padding: 0;
+  border: none;
+  margin: auto 10px;
+  background: none;
+}
 
+.button__img {
+  width: 25px;
+  cursor: pointer;
+}
 
+.covid-table .button__expand {
+  right: 258px;
+  position: absolute;
+  top: 80px;
+
+}
 
 #map {
   width: calc(100vw - 300px);
@@ -162,6 +179,11 @@ h1 {
   background: aliceblu
     /* width: 1500px;
   height: 800px;*/
+}
+
+.map-wrapper {
+
+  background-color: #f5fdff;
 }
 
 .map-wrapper .list-control .search-country {
@@ -220,13 +242,48 @@ h1 {
   /* justify-self: baseline;*/
 }
 
-/*position: absolute;
-top: 0;
-left: 0;
-width: 100%;
-height: 100%;
-z-index: 1000;
-background: aliceblue;*/
+.full-screen {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 1000;
+  margin: 0;
+  padding: 0;
+  /*background: aliceblue;*/
+}
+
+.full-screen #map {
+  width: 100%;
+  height: 100%;
+  /*height: 100%;*/
+  background: aliceblu
+    /* width: 1500px;
+  height: 800px;*/
+}
+
+.full-screen .list-control {
+  right: 25px;
+  top: 20px;
+}
+
+.full-screen #diagram {
+  height: 90%;
+}
+
+.full-screen .covid-list {
+  height: 85%;
+}
+
+.full-screen.covid-table {
+  justify-content: center;
+}
+
+.full-screen.covid-table .button__expand {
+  right: 20px;
+  top: 20px;
+}
 
 .checkbox__one-hundr-thous_label {
   margin-left: 5px;
